+++ src/main/java/blackjack/domain/participant/Name.java
@@ -1,4 +1,4 @@
-package blackjack.participant;
+package blackjack.domain.participant;
 
 
 import utils.StringUtils;
