+++ src/main/java/blackjack/view/InputView.java
@@ -1,0 +1,4 @@
+package blackjack.view;
+
+public class InputView {
+}
