+++ src/main/java/blackjack/view/OutputView.java
@@ -1,0 +1,4 @@
+package blackjack.view;
+
+public class OutputView {
+}
