+++ src/test/java/blackjack/domain/GameTest.java
@@ -1,8 +1,9 @@
-package blackjack;
+package blackjack.domain;
 
-import blackjack.participant.Dealer;
-import blackjack.participant.Participant;
-import blackjack.participant.Player;
+import blackjack.domain.Game;
+import blackjack.domain.participant.Dealer;
+import blackjack.domain.participant.Participant;
+import blackjack.domain.participant.Player;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
