+++ src/test/java/blackjack/domain/participant/NameTest.java
@@ -1,6 +1,5 @@
-package blackjack.participant;
+package blackjack.domain.participant;
 
-import blackjack.participant.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
